Change type of maxTimeToLive in Options struct from uint16_t to uint8_t

+++ include/wstr.h
@@ -16,7 +16,7 @@ struct Options {
     char* destinationHost;
     char* interface;
     uint8_t fqdnFlag;
-    uint16_t maxTimeToLive;
+    uint8_t maxTimeToLive;
 };
 
 /**
